backend/requirements.txt
flask
flask-cors
flask-mail
python-dotenv
